update react-scripts and fix breaking changes
