+++ apps/client/package.json
@@ -18,6 +18,9 @@
     "@types/node": "^12.19.3",
     "@types/react": "^17.0.39",
     "@types/react-dom": "^17.0.11",
+    "@types/file-saver": "^2.0.1",
+    "@types/react-router-dom": "^5.1.6",
+    "@types/uuid": "^8.3.0",
     "@types/react-router": "^5.1.8",
     "clsx": "^1.1.1",
     "file-saver": "^2.0.2",
@@ -27,14 +30,14 @@
     "notistack": "^1.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-hook-form": "^6.11.0",
     "react-full-screen": "^0.3.1",
     "react-json-view": "^1.21.3",
     "react-refresh": "^0.12.0",
     "react-router": "^5.1.8",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.0",
     "typescript": "4.0.3",
-    "web-vitals": "^0.2.4"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -65,10 +68,7 @@
     ]
   },
   "devDependencies": {
-    "@types/file-saver": "^2.0.1",
-    "@types/react-router-dom": "^5.1.6",
-    "@types/uuid": "^8.3.0",
-    "react-hook-form": "^6.11.0"
+    "react-scripts": "5.0.0"
   },
   "proxy": "http://scoretrak.example.com:30080"
 }
